html/semantics/embedded-content-0/the-embed-element/embed-dimension.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTML Test: dimension</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://www.w3.org/html/wg/drafts/html/CR/embedded-content-0.html#the-embed-element">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<embed src="/images/blue.png" height="100" width="100" id="test">
<script>
  test(function () {
    var embed = document.getElementById("test"),
        h = getComputedStyle(embed, false)["height"];       
    assert_equals(h, "100px", "The height of the embed element should be 100px.");
  }, "Check the actual length of the embed element's height");

  test(function () {
    var embed = document.getElementById("test"),
        w = getComputedStyle(embed, false)["width"];
    assert_equals(w, "100px", "The width of the embed element should be 100px.");
  }, "Check the actual length of the embed element's width");
</script>
